tratamento de erro com 404

+++ forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -27,6 +27,7 @@ import br.com.alura.forum.repository.CursoRepository;
 import br.com.alura.forum.repository.TopicoRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -66,30 +67,47 @@ public class TopicosController {
 	}
 	
 	@GetMapping("/{id}")
-	public DetalhesTopicoDto detalhar(@PathVariable Long id) {
+	public ResponseEntity<DetalhesTopicoDto> detalhar(@PathVariable Long id) {
 		
-		Topico topico = topicoRepository.getOne(id);
+		Optional<Topico> topico = topicoRepository.findById(id);
 		
-		return new DetalhesTopicoDto(topico);
+		if (topico.isPresent()) {
+			return ResponseEntity.ok(new DetalhesTopicoDto(topico.get()));	
+		}
+		
+		return ResponseEntity.notFound().build();
 		
 	}
 	
 	@PutMapping("/{id}")
 	@Transactional //garante que o banco de dados seja atualizado
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form, UriComponentsBuilder uriBuilder){
-		Topico topico = form.atualizar(id, topicoRepository);
 		
-		return ResponseEntity.ok(new TopicoDto(topico));
+		Optional<Topico> optional = topicoRepository.findById(id);
+		
+		if (optional.isPresent()) {
+			Topico topico = form.atualizar(id, topicoRepository);
+			return ResponseEntity.ok(new TopicoDto(topico));	
+		}
+		
+		return ResponseEntity.notFound().build();
+		
+		
 	}
 	
 	//ResponseEntity<?> é porque eu não sei qual é o tipo
 	@DeleteMapping("/{id}")
 	@Transactional //garante que o banco de dados seja atualizado
-	public ResponseEntity<?> remover(@PathVariable Long id){
+	public ResponseEntity<?> remover(@PathVariable Long id){	
 		
-		topicoRepository.deleteById(id);
+		Optional<Topico> optional = topicoRepository.findById(id);
 		
-		return ResponseEntity.ok().build(); 		
+		if (optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build(); 	
+		}
+		
+		return ResponseEntity.notFound().build();		
 	}
 	
 }
